changed byte count to size_t

+++ src/files/ecs/MemoryManager.hpp
@@ -5,7 +5,7 @@
 struct Memory_pool
 {
   void *m_runTimeData;
-  uint64_t m_bytesAllocated;
+  size_t m_bytesAllocated;
   
   bool m_MemoryActive = false;
